requirements.txt
google-api-python-client
google-auth
requests
jsonschema
